Trying to fix crashes -> still crashing :/

+++ src/me/anno/traffic/Collision.kt
@@ -1,7 +1,6 @@
 package me.anno.traffic
 
-import org.joml.Vector3d
 import org.joml.Vector3f
 
-data class Collision(val mtvAxis: Vector3d, val minOverlap: Double, val relCenter: Vector3f)
+data class Collision(val mtvAxis: Vector3f, val minOverlap: Float, val relCenter: Vector3f)
